feat(layer): add `layer` subaccessor
Add `layer` subaccessor for depth-related points and move such methods
from `point` to `layer`.
