+++ docs/api-reference/layer.rst
@@ -1,0 +1,18 @@
+.. currentmodule:: pandas
+
+Layer Subaccessor
+=================
+
+.. autoaccessorcallable:: DataFrame.geotech.layer
+
+Methods
+-------
+
+.. autosummary::
+   :toctree: api/
+   :template: autosummary/accessor_method.rst
+
+   DataFrame.geotech.layer.get_top
+   DataFrame.geotech.layer.get_center
+   DataFrame.geotech.layer.get_thickness
+   DataFrame.geotech.layer.split_at
