+++ docs/api-reference/point.rst
@@ -22,7 +22,3 @@ Methods
    :template: autosummary/accessor_method.rst
 
    DataFrame.geotech.point.get_group
-   DataFrame.geotech.point.get_top
-   DataFrame.geotech.point.get_center
-   DataFrame.geotech.point.get_thickness
-   DataFrame.geotech.point.split_at_depth
