+++ src/geotech_pandas/accessor.py
@@ -3,6 +3,7 @@
 import pandas as pd
 
 from geotech_pandas.base import GeotechPandasBase
+from geotech_pandas.layer import LayerDataFrameAccessor
 from geotech_pandas.point import PointDataFrameAccessor
 from geotech_pandas.utils import SubAccessor
 
@@ -14,6 +15,7 @@ class GeotechDataFrameAccessor(GeotechPandasBase):
     """  # noqa: D205
 
     point = SubAccessor(PointDataFrameAccessor)
+    layer = SubAccessor(LayerDataFrameAccessor)
 
     def __init__(self, df: pd.DataFrame):
         self._obj = df
